[#11] rename : 익셉션 핸들러 클래스 이름 수정 및 model 패키지 생성
