+++ src/main/java/com/hyungil/productservice/domain/model/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.hyungil.productservice.domain;
+package com.hyungil.productservice.domain.model;
 
 import java.time.LocalDateTime;
 import javax.persistence.Column;
